use MongoRepository as DAO
