--- src/main/java/io/penguinstats/dao/BaseDao.java
@@ -1,11 +1,0 @@
-package io.penguinstats.dao;
-
-import java.util.List;
-
-public interface BaseDao<T> {
-
-	void save(T t);
-
-	List<T> findAll();
-
-}
