+++ src/main/java/io/penguinstats/dao/LimitationDao.java
@@ -1,13 +1,15 @@
 package io.penguinstats.dao;
 
 import io.penguinstats.model.Limitation;
+import org.springframework.data.mongodb.repository.DeleteQuery;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
-public interface LimitationDao extends BaseDao<Limitation> {
+@Repository
+public interface LimitationDao extends MongoRepository<Limitation, String> {
 
+	@DeleteQuery("{'name' : ?0 }")
 	void removeLimitation(String name);
 
-	void updateLimitation(Limitation limitation);
-
 	Limitation findLimitationByName(String name);
-
 }
