+++ src/main/java/io/penguinstats/dao/ZoneDao.java
@@ -1,13 +1,10 @@
 package io.penguinstats.dao;
 
 import io.penguinstats.model.Zone;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
-public interface ZoneDao extends BaseDao<Zone> {
-
-	void removeZone(String zoneId);
-
-	void updateZone(Zone zone);
-
-	Zone findZoneByZoneId(String zoneId);
+@Repository
+public interface ZoneDao extends MongoRepository<Zone, String> {
 
 }
