+++ src/main/java/io/penguinstats/service/DropMatrixServiceImpl.java
@@ -1,15 +1,14 @@
 package io.penguinstats.service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import io.penguinstats.dao.DropMatrixDao;
 import io.penguinstats.model.DropMatrix;
 import io.penguinstats.model.Stage;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
 
 @Service("dropMatrixService")
 public class DropMatrixServiceImpl implements DropMatrixService {
@@ -35,31 +34,35 @@ public class DropMatrixServiceImpl implements DropMatrixService {
 	 * @return void
 	 */
 	public void clearAndUpdateAll(List<DropMatrix> elements) {
-		dropMatrixDao.removeAll();
-		dropMatrixDao.batchSave(elements);
+		dropMatrixDao.deleteAll();
+		dropMatrixDao.saveAll(elements);
 	}
 
 	/**
-	 * @Title: increateQuantityForOneElement
+	 * @Title: increaseQuantityForOneElement
 	 * @Description: Increase quantity for one element.
 	 * @param stageId
 	 * @param itemId
 	 * @param quantity
 	 * @return void
 	 */
-	public void increateQuantityForOneElement(String stageId, String itemId, Integer quantity) {
-		dropMatrixDao.increateQuantityForOneElement(stageId, itemId, quantity);
+	public void increaseQuantityForOneElement(String stageId, String itemId, Integer quantity) {
+		DropMatrix dropMatrix = dropMatrixDao.findDropMatrixByStageIdAndItemId(stageId, itemId);
+		dropMatrix.increaseQuantity(quantity);
+		dropMatrixDao.save(dropMatrix);
 	}
 
 	/**
-	 * @Title: increateTimesForOneStage
+	 * @Title: increaseTimesForOneStage
 	 * @Description: Increase times for all records in the given stage.
 	 * @param stageId
 	 * @param times
 	 * @return void
 	 */
-	public void increateTimesForOneStage(String stageId, Integer times) {
-		dropMatrixDao.increateTimesForOneStage(stageId, times);
+	public void increaseTimesForOneStage(String stageId, Integer times) {
+		List<DropMatrix> dropMatrixList = dropMatrixDao.findDropMatrixByStageId(stageId);
+		dropMatrixList.forEach(e -> e.increaseTimes(times));
+		dropMatrixDao.saveAll(dropMatrixList);
 	}
 
 	/**
@@ -86,7 +89,7 @@ public class DropMatrixServiceImpl implements DropMatrixService {
 		List<DropMatrix> list = new ArrayList<>();
 		for (String itemId : dropSet)
 			list.add(new DropMatrix(stageId, itemId, 0, 0));
-		dropMatrixDao.batchSave(list);
+		dropMatrixDao.saveAll(list);
 		return true;
 	}
